Created POJO for room that includes lists of IngredientEntity and StepEntity objects
Coded this new POJO into RecipesDAO and the database util class

+++ app/src/main/java/com/cheadtech/example/bakingtime/database/IngredientsEntity.java
@@ -14,7 +14,7 @@ public class IngredientsEntity {
     public Integer recipeId;
 
     @ColumnInfo(name = "quantity")
-    public Double quantity;
+    public Float quantity;
 
     @ColumnInfo(name = "measure")
     public String measure;
